% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_simulation.R
\name{find_ancestor_recursive}
\alias{find_ancestor_recursive}
\title{Find ancestor by recursive method}
\usage{
find_ancestor_recursive(tree, data)
}
\arguments{
\item{tree}{a tree of ape class phylo}

\item{data}{the data associated with the tree}
}
\value{
The recursive estimates of all internal nodes of the tree
}
\description{
Find ancestor by recursive method
}
