% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_simulation.R
\name{pairwise_distance}
\alias{pairwise_distance}
\title{Find pairwise distances between tips}
\usage{
pairwise_distance(tr)
}
\arguments{
\item{tr}{a tree of ape class phylo (must be recapitated)}
}
\value{
Pairwise geographic and genetic distances
}
\description{
Find pairwise distances between tips
}
